[B] Improve unauthenticated authorization
This commit addresses a couple issues:

1. We weren't previously authorizing Project/Text#show, which meant
that direct navigation to draft projects and those projects' texts
always succeeded, even when unauthenticated.

2. We were rescuing Authority::MissingUser errors with a 401 error
requiring login.  This is incorrect when an unauthenticated user is
attempting to view a Project or Text and we authorize the action.
Instead, we now create an `anonymous_user` that has reader privileges
and defers its `can_read?` check to the resource we're verifying.

Fixes #1670